pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TeamControl"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "matplotlib",
    "protobuf",
    "PyYAML",
    "py_trees",
    "Shapely",
    "scikit-learn",
    "networkx",
    "pygame",
    "PySide6",
    # behaviour_tree is included locally in this repo for testing
]

dynamic = [
    "version"
]

[tools.setuptools.packages.find]
where = ["src"]

[tools.setuptools.dynamic]
version = {attr = "TeamControl.__version__"}

[project.optional-dependencies]
testing = [
    "pytest",
]